backend/requirements.txt
pyTelegramBotAPI
requests
google-generativeai
openai
fastapi
uvicorn
python-multipart
python-dotenv
sqlalchemy
psycopg2-binary
pgvector
supabase
